Revert "Merge remote-tracking branch 'origin/position_feature' into job-posting"
This reverts commit 8f6f2e49c52bb548c6fc8814e8f41ded3f7ba6b7, reversing
changes made to b4396d80d59385373c1c28d9a499133bb5c1a92f.

+++ db/src/entities/Position.ts
@@ -15,11 +15,12 @@ import { Referral } from './Referral';
 @Entity('position', { schema: 'aki' })
 export class Position {
   // Matt Cappucci - made this auto increment
-  @PrimaryGeneratedColumn({ type: 'int', name: 'id' })
+  @PrimaryGeneratedColumn({ type: 'bigint', name: 'id' })
   id: number;
 
   /*@Column("bigint", { primary: true, name: "id" })
   id: string;*/
+
   @Column('longtext', { name: 'description', nullable: true })
   description: string | null;
